Restructure cli to prepare for providers
