+++ infrastructure/cli/pkg/cmd/BUILD
@@ -4,18 +4,18 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "cmd",
     srcs = ["root.go"],
-    importpath = "cli/cmd",
+    importpath = "cli/pkg/cmd",
     visibility = ["//visibility:public"],
     x_defs = {
         "Version": "{STABLE_VERSION}",
         "CommitSHA1": "{STABLE_GIT_COMMIT}",
     },
     deps = [
-        "//infrastructure/cli/cmd/api",
-        "//infrastructure/cli/cmd/config",
-        "//infrastructure/cli/cmd/create",
-        "//infrastructure/cli/cmd/status",
-        "//infrastructure/cli/cmd/ui",
+        "//infrastructure/cli/pkg/cmd/api",
+        "//infrastructure/cli/pkg/cmd/config",
+        "//infrastructure/cli/pkg/cmd/create",
+        "//infrastructure/cli/pkg/cmd/status",
+        "//infrastructure/cli/pkg/cmd/ui",
         "@com_github_mitchellh_go_homedir//:go-homedir",
         "@com_github_spf13_cobra//:cobra",
         "@com_github_spf13_viper//:viper",
