+++ infrastructure/cli/pkg/cmd/api/signup.go
