+++ infrastructure/cli/pkg/cmd/status/status.go
