+++ infrastructure/cli/pkg/cmd/ui/BUILD
@@ -4,7 +4,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "ui",
     srcs = ["ui.go"],
-    importpath = "cli/cmd/ui",
+    importpath = "cli/pkg/cmd/ui",
     visibility = ["//visibility:public"],
     deps = ["@com_github_spf13_cobra//:cobra"],
 )
